+++ test/bin/builder/inflater_builder_test.dart
@@ -8,9 +8,9 @@ void main() {
 
   test('Test InflaterBuilder dependency resolution', () async {
     final config = BuilderConfig();
+    await config.loadConfig("xwidget|res/default_config.yaml");
     config.inflaterConfig.sources.add("xwidget|test/fixtures/src/inflater_spec.dart");
     config.inflaterConfig.target = "build/test.dart";
-    config.schemaConfig.template = "xwidget|res/schema_template.xsd";
     config.schemaConfig.target = "build/test.xsd";
 
     final builder = InflaterBuilder(config);
@@ -20,9 +20,21 @@ void main() {
 
   test('Test new inflater spec format', () async {
     final config = BuilderConfig();
+    await config.loadConfig("xwidget|res/default_config.yaml");
     config.inflaterConfig.sources.add("xwidget|test/fixtures/src/inflater_spec.dart");
     config.inflaterConfig.target = "build/inflaters_test.g.dart";
-    config.schemaConfig.template = "xwidget|res/schema_template.xsd";
+    config.schemaConfig.target = "build/schema_test.xsd";
+
+    final builder = InflaterBuilder(config);
+    final result = await builder.build();
+    expect(result.outputs.length, 2);
+  });
+
+  test('Test inflater spec generics', () async {
+    final config = BuilderConfig();
+    await config.loadConfig("xwidget|res/default_config.yaml");
+    config.inflaterConfig.sources.add("xwidget|test/fixtures/src/inflater_spec.dart");
+    config.inflaterConfig.target = "build/inflaters_test.g.dart";
     config.schemaConfig.target = "build/schema_test.xsd";
 
     final builder = InflaterBuilder(config);